discogs_credits: extract IDB storage + edit-note builder
Issue #32 — two small modules out:

  src/storage.js
    IDB init (the lazy indexedDB.open('mblink') + mblinks upgrade
    handler), `readIdbRecord(key)`. `db` is exported as a live binding
    (`export let db = null`) so the 15+ direct `db.transaction(...)`
    reads still scattered through the entry see the IDBDatabase once
    open — no need to rewrite every call site.

  src/edit-note.js
    `buildEditNote(discogsUrl, opts, extraLines)`. Pure render of the
    "this edit was made by …" preamble + URLs + options summary that
    drops into MB's edit-note input at import start. Docstring revised
    to remove the dead reference to `updateSummary` (deleted last
    chunk).

~60 lines moved out. No behavior change; full 12-fixture suite green.
