+++ userscripts/discogs_credits/dev/DECISIONS.md
@@ -82,4 +82,5 @@ For the full reasoning behind any decision, `git log -- dev/ANALYSIS.md` finds t
 - 2026-05-24 22:50 — issue #32 chunk 2e (logger + resolveLinkTypeId) → extracted `addLogLine` to `src/log.js`. The log container (the `<ul>` element the UI bar creates at startup) is now wired in via `setLogContainer(el)` after the UI bar constructs it; this lets log.js be a standalone module without circular access to the entry's `let logs`. Pre-init `addLogLine` calls silently no-op instead of throwing (slightly safer than today). With the logger out, `resolveLinkTypeId` follows it into `src/api-mb.js` (it depends on `addLogLine`). Entry no longer has either function definition. Smoke fixture passes; full suite running.
 - 2026-05-24 22:56 — issue #32 proposal C (dead-code removal) → deleted the legacy DOM-automation dispatch chain in one pass: `addRelationshipsForCompanies` / `addRelationshipsForArtists` / `addRelationshipsForTracklist` (callers, never reached today), `addReleaseRelationship` / `addRelationship` (the DOM-driven dispatch they call), `updateSummary` (only called from `addRelationship`), `getElementByLabel` / `checkAdditionalAttribute` (helpers of `addRelationship`), `getMbId` / `scheduleRequest` (the 300 ms-gap request chain only `addRelationshipsFor*` used — superseded by `mbThrottle`), `getTrackRowBasedOnTrack` (only used by `addRelationshipsForTracklist`), `setValueOnAutocomplete` (only used by `addRelationship`), and the module-level `let lastRequest` / `let lastUiItem` (only written/read by the deleted chain). ~410 lines removed from the entry. Bundle output size unchanged — esbuild was already tree-shaking these. Full suite still green.
 - 2026-05-24 23:05 — issue #32 chunk 2f (mappers) → extracted the five Discogs↔role transforms to `src/mappers.js`: `guessSortName`, `getAllArtistTracks`, `convertPotentialDJMixers`, `getArtistRoles`, `convertDiscogsArtistsToRolesRelationships`. All pure (or only mutate their direct inputs); deps are just `ENTITY_TYPE_MAP`, `INSTRUMENTS`, `SELECTORS` — all already extracted. ~280 lines moved out. Smoke fixture passes; full suite running.
+- 2026-05-24 23:11 — issue #32 chunk 2g (storage + edit-note) → extracted IDB init (the lazy `indexedDB.open('mblink')` block + the `mblinks` upgrade handler) to `src/storage.js` along with `readIdbRecord`. `db` is exported as a *live binding* (`export let db = null`) so the 15+ direct `db.transaction(...)` reads still scattered through the entry see the value once IDB finishes opening — no rewrite required. `buildEditNote` moved to `src/edit-note.js` (small standalone, GM_info-shaped). Updated its docstring (the dead `updateSummary` reference is gone). ~35 + ~25 lines out. Smoke passes; full suite running.
 - 2026-05-24 19:55 — issue #30 (entity-fetch slower than the old "5-thread burst" approach) → `mbThrottle` rewritten. The previous shape was a strict serial promise chain with a post-success `setTimeout(_gap)` between each request (`_gap` ∈ [200, 3000] ms), so the 5-worker pools in `checkMissingArtists` / `checkMissingCompanies` all funnelled through one queue at ≤5 req/s — slower than the old "5 immediate parallel threads" approach the throttle was meant to replace. New shape: up to `MAX_CONCURRENT=4` in-flight requests, no artificial gap, **global cooperative `_pauseUntil`** on 429/503 that honors `Retry-After` (or exponential per-request backoff if absent). All in-flight workers idle together until the pause elapses, which avoids the thundering-herd retries that the original 5-thread approach produced. Also removed two redundant `setTimeout(200)` calls (showReviewTable name pre-load loop and the URL-check pool) — the throttle paces. Expected ~3–5× speedup on rate-limited-heavy fixtures (Midwest Funk-style preflight), without 503 storms.
